pyproject.toml
[project]
name = "octobatch"
version = "0.1.0"
description = "Foundational control plane services for Octobatch"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Octobatch Team" }]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "httpx>=0.26.0",
    "redis[hiredis]>=5.0.0",
    "asyncpg>=0.29.0",
    "aioboto3>=12.4.0",
    "PyJWT>=2.8.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
octobatch-api-gateway = "octobatch.api_gateway.main:main"
octobatch-campaign = "octobatch.campaign.main:main"
octobatch-selection = "octobatch.selection.main:main"
octobatch-token = "octobatch.token.main:main"
